weather_forecast/requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
djangorestframework==3.14.0
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.3.0
pydantic_core==2.6.3
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.6
yarl==1.9.2
